fixing code smells

+++ src/dsp/doppler.c
@@ -40,7 +40,7 @@ int32_t doppler_calculate_shift(doppler *result, int direction) {
     return direction * (result->center_freq - result->center_freq * (SPEED_OF_LIGHT - result->obs_set->range_rate) / SPEED_OF_LIGHT) + result->constant_offset;
 }
 
-int doppler_create(float latitude, float longitude, float altitude, uint32_t sampling_freq, uint32_t center_freq, int32_t constant_offset, time_t start_time_seconds, uint32_t max_output_buffer_length, char tle[3][80], doppler **d) {
+int doppler_create(double latitude, double longitude, double altitude, uint32_t sampling_freq, uint32_t center_freq, int32_t constant_offset, time_t start_time_seconds, uint32_t max_output_buffer_length, char tle[3][80], doppler **d) {
     struct doppler_t *result = malloc(sizeof(struct doppler_t));
     if (result == NULL) {
         return -ENOMEM;
